// frontend/src/app/generated/model/modelsSitePoolEntry.ts
/**
 * Homelab API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ModelsSitePoolEntry { 
    tags?: Array<string>;
    /**
     * 0:Keyword, 1:Regex, 2:Domain, 3:Full
     */
    type?: number;
    value?: string;
}